refactor(useHashIdScroll): Add logic of scroll to hashId in hook
